implementing story 1

+++ backend/domain/fornecedor/src/main/java/com/cesar/school/fds2/raycharge/fornecedor/domain/estacaoderecarga/EstacaoDeRecarga.java
@@ -17,13 +17,13 @@ public class EstacaoDeRecarga implements Cloneable, AggregateRoot<EstacaoDeRecar
   private StatusEstacao statusEstacao;
   private int precoMinimo;
   private int precoPKwH;
-  private int distancia;
+  private float distancia;
   private int tempoPorAgendamento;
   private List<HorarioDisponivel> horarioDisponiveis;
   private List<IdAgendamento> historicoDeUso;
 
 
-  public EstacaoDeRecarga(IdEstacao idEstacao, IdFornecedor idFornecedor, String nomeDaEstacao, int quantidadeDeCarregadores, HorarioDisponivel horarioFuncionamento, Endereco enderecoEstacao, StatusEstacao statusEstacao, int precoMinimo, int precoPKwH, int distancia, int tempoPorAgendamento, List<HorarioDisponivel> horarioDisponiveis, List<IdAgendamento> historicoDeUso) {
+  public EstacaoDeRecarga(IdEstacao idEstacao, IdFornecedor idFornecedor, String nomeDaEstacao, int quantidadeDeCarregadores, HorarioDisponivel horarioFuncionamento, Endereco enderecoEstacao, StatusEstacao statusEstacao, int precoMinimo, int precoPKwH, float distancia, int tempoPorAgendamento, List<HorarioDisponivel> horarioDisponiveis, List<IdAgendamento> historicoDeUso) {
     this.idEstacao = idEstacao;
     this.idFornecedor = idFornecedor;
     this.nomeDaEstacao = nomeDaEstacao;
@@ -33,7 +33,7 @@ public class EstacaoDeRecarga implements Cloneable, AggregateRoot<EstacaoDeRecar
     this.statusEstacao = statusEstacao;
     this.precoMinimo = precoMinimo;
     this.precoPKwH = precoPKwH;
-    this.distancia = 0; // distancia como NULL
+    this.distancia = 0.0f; // distancia como NULL
     this.tempoPorAgendamento = tempoPorAgendamento;
     this.horarioDisponiveis = horarioDisponiveis;
     this.historicoDeUso = historicoDeUso;
@@ -104,11 +104,11 @@ public class EstacaoDeRecarga implements Cloneable, AggregateRoot<EstacaoDeRecar
     this.precoPKwH = precoPKwH;
   }
 
-  public int getDistancia() {
+  public float getDistancia() {
     return distancia;
   }
 
-  public void setDistancia(int distancia) {
+  public void setDistancia(float distancia) {
     this.distancia = distancia;
   }
